v2 of DICOM Structured Content Anonymization Library

+++ src/main/java/eu/interopehrate/dicom_anonymization_library/DICOMAnonymization.java
@@ -6,10 +6,17 @@ import org.json.JSONObject;
 import java.io.File;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.HashMap;
 
 import static org.toilelibre.libe.curl.Curl.curl;
 
 public class DICOMAnonymization implements DICOMAnonymizationI {
+
+    private static final HashMap<String, String> baseUrl = new HashMap<>() {{
+        put("CHU", "http://139.165.99.12:8000/uploadfile");
+        put("FTGM", "http://10.97.32.223:9000/uploadfile");
+        put("LOCALHOST", "http://localhost:8000/uploadfile");
+    }};
 
     private String declareInput(String filename) {
         String input = "";
@@ -55,29 +62,45 @@ public class DICOMAnonymization implements DICOMAnonymizationI {
         }*/
     }
 
-    public String invokeEndpoint(String filename) {
+    private boolean healthcareOrganizationExists(String healthcareOrganization) {
+        if(baseUrl.containsKey(healthcareOrganization)) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    public String invokeEndpoint(String filename, String healthcareOrganization) {
 
         String input = declareInput(filename);
         String filepath = declareFilepath(filename);
         String output = declareOutput(input);
-        String url = "http://localhost:8000/uploadfile";
-        String command = "curl -X POST '" + url + "' -F 'dicom=@" + filepath + input + "' -o " + filepath + output;
-        //System.out.println("Command: " + command);
-
-        HttpResponse response = curl(command);
-        int status = response.getStatusLine().getStatusCode();
 
         JSONObject responseMessage = new JSONObject();
-        responseMessage.put("status", status);
         String message = "";
-        if (status == 200) {
-            message = "File '" + input + "' was successfully anonymized.";
-        } else if (status == 400) {
-            message = "Anonymization process was not applied to this file. File '" + input + "' is NOT in DICOM format.";
-            deleteFile(filepath + output);
+
+        if(healthcareOrganizationExists(healthcareOrganization)) {
+            String url = baseUrl.get(healthcareOrganization);
+            String command = "curl -X POST '" + url + "' -F 'dicom=@" + filepath + input + "' -o " + filepath + output;
+            //System.out.println("Command: " + command);
+
+            HttpResponse response = curl(command);
+            int status = response.getStatusLine().getStatusCode();
+
+            responseMessage.put("status", status);
+
+            if (status == 200) {
+                message = "File '" + input + "' was successfully anonymized.";
+            } else if (status == 400) {
+                message = "Anonymization process was not applied to this file. File '" + input + "' is NOT in DICOM format.";
+                deleteFile(filepath + output);
+            } else {
+                message = response.getStatusLine().getReasonPhrase();
+            }
         } else {
-            message = response.getStatusLine().getReasonPhrase();
+            message = "The Healthcare Organization provided does not exist. Please choose one in range: " + baseUrl.keySet().toString();
         }
+
         responseMessage.put("message", message);
 
         return responseMessage.toString();
